correctly logging start and end times
Manually tested so times seem to be roughly correct. Also made sure for each player, they end with an "end" speakerEvent. The only issue is "active-speaker-change" is fired only when a user starts talking, but not when a user stops talking. I haven't figured out a way to detect volume etc. #99