add talkMeter
the best way to test this seems to be to actually display the talk time information as part of a talk meter. we've been meaning to build this anyways, so I just made a basic version.

One thing that we don't account for at the moment is the time before anyone starts talking, in the proportion calculation.