talk timing (#549)
* element time cannot exceed stage duration

* restore original treatments.test.yaml

* log each participant's speak time (hasn't tested)

* logging speaker events start time

* logging start and end time but w errors

time looks weird, but not be accurate

* correctly logging start and end times

Manually tested so times seem to be roughly correct. Also made sure for each player, they end with an "end" speakerEvent. The only issue is "active-speaker-change" is fired only when a user starts talking, but not when a user stops talking. I haven't figured out a way to detect volume etc. #99

* back to log by active speaker mode

* pull out logEndTime()

* stageTimer working

* stash for changes to video call

* add talkMeter

the best way to test this seems to be to actually display the talk time information as part of a talk meter. we've been meaning to build this anyways, so I just made a basic version.

One thing that we don't account for at the moment is the time before anyone starts talking, in the proportion calculation.

* move away from broken append

---------

Co-authored-by: gumillie222 <121700352+gumillie222@users.noreply.github.com>

closes: https://github.com/Watts-Lab/deliberation-empirica/issues/99